requirements.txt
anthropic
openai
pymupdf
python-dotenv
tiktoken
